Try cleaning up and reusing more code in crc and handle full packet data (trimming only to correct length and not skipping any header data)

+++ src/embodycodec/codec.py
@@ -19,7 +19,7 @@ from typing import Union
 from embodycodec import attributes as a
 from embodycodec import types as t
 from embodycodec.crc import crc16
-from embodycodec.exceptions import DecodeError
+from embodycodec.exceptions import DecodeError,CrcError
 
 
 T = TypeVar("T", bound="Message")
@@ -28,6 +28,15 @@ T = TypeVar("T", bound="Message")
 @dataclass
 class Message(ABC):
     """Abstract base class for protocol messages"""
+
+    hdr_len = 3
+    """Header length to avoid checking every time"""
+    
+    crc_len = 2
+    """CRC length to avoid magic numbers"""
+
+    struct_hdr_format = ">BH"
+    """unpack format that is fixed for all headers, taking up hdr_len bytes"""
 
     struct_format = ""
     """unpack format to be overridden by sub-classes, see
@@ -48,29 +57,50 @@ class Message(ABC):
         return struct.calcsize(cls.struct_format)
 
     @classmethod
+    def _check_crc_and_get_length(cls, data: bytes) -> int:
+        (data_type, data_length,) = struct.unpack(cls.struct_hdr_format, data[0:cls.hdr_len])
+        if len(data) < data_length:
+            raise BufferError(
+                f"Buffer too short for message: Received {len(data)} bytes, expected {data_length} bytes"
+            )   # Note: This is not technically an error as more data may arrive allowing the message to be decoded,
+                # but raised as error to split it from the resulting length found in the process of checking crc
+        (crc,) = struct.unpack(">H", data[data_length-2:data_length])
+        calculated_crc = crc16(data[0:data_length-2])
+        if crc != calculated_crc:
+            raise CrcError(
+                f"CRC error: Calculated {calculated_crc:04X}, received {crc:04X}"
+            )
+        return data_length
+
+    @classmethod
     def decode(cls: type[T], data: bytes) -> T:
         """Decode bytes into message object"""
-        pos = 2  # offset to start of body (skips length field)
+        length = cls._check_crc_and_get_length(data)
+        pos = hdr_len  # offset to start of body (skips msg_type and length field)
         msg = cls(
             *(struct.unpack(cls.struct_format, data[pos : pos + cls.__body_length()]))
         )
-        (msg.length,) = struct.unpack(">H", data[0:pos])
-        (msg.crc,) = struct.unpack(
-            ">H", data[pos + cls.__body_length() : msg.length - 1]
-        )
+        msg.length = length
         return msg
 
+    # Note: Default method encodes dataclass with fixed length and format
     def encode(self) -> bytes:
         """Encode a message object to bytes"""
         body = self._encode_body()
-        header = struct.pack(">BH", self.msg_type, len(body) + 5)
+        header = _encode_header(body)
         header_and_body = header + body
-        crc_calculated = crc16(header_and_body)
-        crc = struct.pack(">H", crc_calculated)
-        return header_and_body + crc
+        return header_and_body + _encode_crc(header_and_body)
 
     def _encode_body(self) -> bytes:
         return struct.pack(self.struct_format, *astuple(self))
+
+    def _encode_crc(header_and_body: bytes) -> bytes:
+        crc_calculated = crc16(header_and_body)
+        crc = struct.pack(">H", crc_calculated)
+        return crc
+
+    def _encode_header(body: bytes) -> bytes:
+        return struct.pack(struct_hdr_format, self.msg_type, len(body) + hdr_len + crc_len)
 
 
 @dataclass
@@ -115,12 +145,12 @@ class SetAttribute(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "SetAttribute":
-        pos = 2  # offset to start of body (skips length field)
-        (attribute_id,) = struct.unpack(">B", data[pos : pos + 1])
-        (attrib_len,) = struct.unpack(">B", data[pos + 1 : pos + 2])
+        length = cls._check_crc_and_get_length(data)
+        pos = cls.hdr_len  # offset to start of body (skips msg_type and length field)
+        (attribute_id,attrib_len,) = struct.unpack(">BB", data[pos : pos + 2])
         value = a.decode_attribute(attribute_id, data[pos + 2 : pos + 2 + attrib_len])
         msg = SetAttribute(attribute_id=attribute_id, value=value)
-        (msg.length,) = struct.unpack(">H", data[0:pos])
+        msg.length = length
         return msg
 
     def _encode_body(self) -> bytes:
@@ -152,7 +182,8 @@ class GetAttributeResponse(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "GetAttributeResponse":
-        pos = 2  # offset to start of body (skips length field)
+        data_length = cls._check_crc_and_get_length(data)
+        pos = cls.hdr_len  # offset to start of body (skips msg_type and length field)
         (attribute_id,) = struct.unpack(">B", data[pos + 0 : pos + 1])
         (changed_at,) = struct.unpack(">Q", data[pos + 1 : pos + 9])
         reporting = t.Reporting.decode(
@@ -167,7 +198,7 @@ class GetAttributeResponse(Message):
             reporting=reporting,
             value=value,
         )
-        (msg.length,) = struct.unpack(">H", data[:2])
+        (msg.length,) = data_length
         return msg
 
     def _encode_body(self) -> bytes:
@@ -198,13 +229,14 @@ class ConfigureReporting(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "ConfigureReporting":
-        pos = 2  # offset to start of body (skips length field)
+        length = cls._check_crc_and_get_length(data)
+        pos = cls.hdr_len  # offset to start of body (skips msg_type and length field)
         (attribute_id,) = struct.unpack(">B", data[pos + 0 : pos + 1])
         reporting = t.Reporting.decode(
             data[pos + 1 : pos + 1 + t.Reporting.default_length()]
         )
         msg = ConfigureReporting(attribute_id=attribute_id, reporting=reporting)
-        (msg.length,) = struct.unpack(">H", data[0:pos])
+        (msg.length,) = length
         return msg
 
     def _encode_body(self) -> bytes:
@@ -219,9 +251,10 @@ class ConfigureReportingResponse(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "ConfigureReportingResponse":
-        pos = 2  # offset to start of body (skips length field)
+        if not cls.__check_crc(data): return None
+        pos = 3  # offset to start of body (skips length field)
         msg = ConfigureReportingResponse()
-        (msg.length,) = struct.unpack(">H", data[0:pos])
+        (msg.length,) = struct.unpack(">H", data[1:pos])
         return msg
 
 
@@ -244,12 +277,13 @@ class PeriodicRecording(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "PeriodicRecording":
-        pos = 2  # offset to start of body (skips length field)
+        if not cls.__check_crc(data): return None
+        pos = 3  # offset to start of body (skips length field)
         recording = t.Recording.decode(
             data[pos + 0 : pos + t.Recording.default_length()]
         )
         msg = PeriodicRecording(recording=recording)
-        (msg.length,) = struct.unpack(">H", data[0:pos])
+        (msg.length,) = struct.unpack(">H", data[1:pos])
         return msg
 
     def _encode_body(self) -> bytes:
@@ -270,14 +304,15 @@ class AttributeChanged(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "AttributeChanged":
-        pos = 2  # offset to start of body (skips length field)
+        if not cls.__check_crc(data): return None
+        pos = 3  # offset to start of body (skips length field)
         (changed_at,) = struct.unpack(">Q", data[pos + 0 : pos + 8])
         (attribute_id,) = struct.unpack(">B", data[pos + 8 : pos + 9])
         value = a.decode_attribute(attribute_id, data[pos + 10 :])
         msg = AttributeChanged(
             changed_at=changed_at, attribute_id=attribute_id, value=value
         )
-        (msg.length,) = struct.unpack(">H", data[0:pos])
+        (msg.length,) = struct.unpack(">H", data[1:pos])
         return msg
 
     def _encode_body(self) -> bytes:
@@ -300,10 +335,11 @@ class RawPulseChanged(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "RawPulseChanged":
-        pos = 2  # offset to start of body (skips length field (2B))
+        if not cls.__check_crc(data): return None
+        pos = 3  # offset to start of body (skips type (1B) and length field (2B))
         header_crc = 7  # attrib_id (1B) + length (2B) + changed_at (2B) + crc (2B)
         (changed_at,) = struct.unpack(">H", data[pos + 0 : pos + 2])
-        (length,) = struct.unpack(">H", data[0:pos])
+        (length,) = struct.unpack(">H", data[1:pos])
         # Determine if payload contains 1 or 3 PPGs
         if length - header_crc == t.PulseRawAll.default_length():
             value = t.PulseRawAll.decode(
@@ -334,11 +370,12 @@ class RawPulseListChanged(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "RawPulseListChanged":
-        pos = 2  # offset to start of body (skips length field (2B))
+        if not cls.__check_crc(data): return None
+        pos = 3  # offset to start of body (skips length field (2B))
         (attribute_id,) = struct.unpack(">B", data[pos : pos + 1])
         value = a.PulseRawListAttribute.decode(data[pos + 1 :])
         msg = RawPulseListChanged(attribute_id=attribute_id, value=value)
-        (msg.length,) = struct.unpack(">H", data[0:pos])
+        (msg.length,) = struct.unpack(">H", data[1:pos])
         return msg
 
     def _encode_body(self) -> bytes:
@@ -384,12 +421,13 @@ class ListFilesResponse(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "ListFilesResponse":
+        if not cls.__check_crc(data): return None
         # ListFiles length
         msg = ListFilesResponse(files=[])
         (msg.length,) = struct.unpack(">H", data[0:2])
 
         if msg.length > 5:
-            pos = 2
+            pos = 3
             while pos + t.FileWithLength.default_length() <= msg.length - 1:
                 msg.files.append(
                     t.FileWithLength.decode(
@@ -415,10 +453,11 @@ class GetFile(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "GetFile":
-        pos = 2  # offset to start of body (skips length field)
+        if not cls.__check_crc(data): return None
+        pos = 3  # offset to start of body (skips length field)
         value = t.File.decode(data[pos:])
         msg = GetFile(file=value)
-        (msg.length,) = struct.unpack(">H", data[0:pos])
+        (msg.length,) = struct.unpack(">H", data[1:pos])
         return msg
 
     def _encode_body(self) -> bytes:
@@ -440,7 +479,8 @@ class SendFile(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "SendFile":
-        pos = 2  # offset to start of body (skips length field)
+        if not cls.__check_crc(data): return None
+        pos = 3  # offset to start of body (skips length field)
         file_name = t.File.decode(data[pos:])
         (index,) = struct.unpack(
             ">H",
@@ -454,7 +494,7 @@ class SendFile(Message):
         msg = SendFile(
             file_name=file_name, index=index, total_parts=total_parts, payload=payload
         )
-        (msg.length,) = struct.unpack(">H", data[0:pos])
+        (msg.length,) = struct.unpack(">H", data[1:pos])
         return msg
 
     def _encode_body(self) -> bytes:
@@ -479,10 +519,11 @@ class DeleteFile(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "DeleteFile":
-        pos = 2  # offset to start of body (skips length field)
+        if not cls.__check_crc(data): return None
+        pos = 3  # offset to start of body (skips length field)
         value = t.File.decode(data[pos:])
         msg = DeleteFile(file=value)
-        (msg.length,) = struct.unpack(">H", data[0:pos])
+        (msg.length,) = struct.unpack(">H", data[1:pos])
         return msg
 
     def _encode_body(self) -> bytes:
@@ -501,10 +542,11 @@ class GetFileUart(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "GetFileUart":
-        pos = 2  # offset to start of body (skips length field)
+        if not cls.__check_crc(data): return None
+        pos = 3  # offset to start of body (skips length field)
         value = t.File.decode(data[pos:])
         msg = GetFileUart(file=value)
-        (msg.length,) = struct.unpack(">H", data[0:pos])
+        (msg.length,) = struct.unpack(">H", data[1:pos])
         return msg
 
     def _encode_body(self) -> bytes:
@@ -562,11 +604,12 @@ class ExecuteCommand(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "ExecuteCommand":
-        pos = 2  # offset to start of body (skips length field)
+        if not cls.__check_crc(data): return None
+        pos = 3  # offset to start of body (skips length field)
         (command_id,) = struct.unpack(">B", data[pos + 0 : pos + 1])
         value = data
         msg = ExecuteCommand(command_id=command_id, value=value)
-        (msg.length,) = struct.unpack(">H", data[0:pos])
+        (msg.length,) = struct.unpack(">H", data[1:pos])
         return msg
 
     def _encode_body(self) -> bytes:
@@ -622,11 +665,12 @@ class ExecuteCommandResponse(Message):
 
     @classmethod
     def decode(cls, data: bytes) -> "ExecuteCommandResponse":
-        pos = 2  # offset to start of body (skips length field)
+        if not cls.__check_crc(data): return None
+        pos = 3  # offset to start of body (skips length field)
         (response_code,) = struct.unpack(">B", data[pos + 0 : pos + 1])
         value = data
         msg = ExecuteCommandResponse(response_code=response_code, value=value)
-        (msg.length,) = struct.unpack(">H", data[0:pos])
+        (msg.length,) = struct.unpack(">H", data[1:pos])
         return msg
 
     def _encode_body(self) -> bytes:
@@ -649,17 +693,7 @@ def decode(data: bytes) -> Message:
     """
 
     (message_type, length,) = struct.unpack(">BH", data[0:3])
-    if len(data) < length:
-        raise BufferError(
-            f"Buffer too short for message: Received {len(data)} bytes, expected {length} bytes"
-        ) # Note: This is not technically an error as more data may arrive allowing the message to be decoded
-    (crc,) = struct.unpack(">H", data[length-2:length])
-    calculated_crc = crc16(data[0:length-2])
-    if crc != calculated_crc:
-        raise CrcError(
-            f"CRC error: Calculated {calculated_crc:04X}, received {crc:04X}"
-        ) # Note: This is not technically an error as more data may arrive allowing the message to be decoded
-    trimmed_data = data[3:length-2] # Prepare the data by trimming off the message type, length and crc field
+    trimmed_data = data[0:length] # Prepare the data by trimming off any additional data not part of packet
     try:
         if message_type == Heartbeat.msg_type:
             return Heartbeat.decode(trimmed_data)
